refactor code to solve linters
