--- Person.rb
@@ -1,24 +1,0 @@
-require './Nameable'
-class Person < Nameable
-    attr_reader :id, :parent_permission
-    attr_accessor :name, :age
-  
-    def initialize(age, name = 'Unknown', parent_permission: true)
-      @id = Random.rand(1..1000)
-      @name = name
-      @age = age
-      @parent_permission = parent_permission
-    end
-  
-    private
-  
-    def of_age?
-      @age >= 18
-    end
-  
-    public
-  
-    def can_use_services?
-      of_age? || @parent_permission
-    end
-  end
