+++ nameable.rb
@@ -1,0 +1,5 @@
+class Nameable
+  def correct_name()
+    raise NotImplementedError, "#{self.class} has not implemented method '#{__method__}'"
+  end
+end
